shared/common/pyproject.toml
[tool.poetry]
name = "common"
version = "0.1.0"
description = "Shared common library for all services"
authors = ["AI Middle Team"]
readme = "README.md"
packages = [{include = "common"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5.3"
structlog = "^24.1.0"
prometheus-client = "^0.19.0"
opentelemetry-api = "^1.22.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
black = "^23.12.1"
ruff = "^0.1.9"
mypy = "^1.8.0"

[tool.pytest.ini_options]
minversion = "7.0"
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = ["--cov=common", "--cov-report=term-missing"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "W", "F", "I", "C", "B", "UP", "N", "S", "T20", "PT", "RET", "SIM", "ARG"]
ignore = ["S101", "B008"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

